+++ primitives/arkworks/curves/bls12_381/Cargo.toml
@@ -17,7 +17,7 @@ ark-ff = { version="^0.3.0", default-features = false }
 ark-ec = { version="^0.3.0", default-features = false }
 ark-std = { version="^0.3.0", default-features = false }
 ark-serialize = { version="^0.3.0", default-features = false }
-bls12 = { package = "sub-bls12-model", version = "^0.3.0", path = "../../models/bls12", default-features = false }
+ark-sub-models = { package = "ark-sub-models", version = "^0.3.0", path = "../../models", default-features = false }
 ark-bls12-381 = { git = "https://github.com/arkworks-rs/curves/" }
 sp-io = { version = "6.0.0", default-features = false, path = "../../../io" }
 
@@ -27,7 +27,7 @@ hex = "^0.4.0"
 
 [features]
 default = [ "curve" ]
-std = [ "ark-std/std", "ark-ff/std", "ark-ec/std", "bls12/std", "sp-io/std" ]
+std = [ "ark-std/std", "ark-ff/std", "ark-ec/std", "ark-sub-models/std", "sp-io/std" ]
 
 curve = [ "scalar_field" ]
 scalar_field = []
