+++ primitives/arkworks/models/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sub-bls12-model"
+name = "ark-sub-models"
 version = "0.3.0"
 authors = [ "arkworks contributors" ]
 description = "A substrate replacement model for bls12 of ark-ec"
@@ -24,7 +24,6 @@ getrandom = { version = "0.2", features = ["js"] }
 rayon = { version = "1", optional = true }
 zeroize = { version = "1", default-features = false, features = ["zeroize_derive"] }
 itertools = { version = "0.10", default-features = false }
-short-weierstrass = { package = "sub-short-weierstrass-model", version = "^0.3.0", path = "../short_weierstrass", default-features = false }
 
 [features]
 default = [ ]
