+++ primitives/arkworks/models/src/lib.rs
@@ -1,0 +1,3 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod bls12;
